+++ ui/src/hooks/index.ts
@@ -9,3 +9,4 @@ export * from './useProofResult'
 export * from './useResetAnonSet'
 export * from './useStore'
 export * from './useVerifyProof'
+export * from './useWorker'
