+++ ui/src/hooks/useMerkleTreeWorker.ts
@@ -2,7 +2,7 @@ import {
   GenerateMerkleProofFn,
   MerkleTreeWorker,
 } from '@anonklub/merkle-tree-worker'
-import { useWorker } from '@/hooks/useWorker'
+import { useWorker } from '@/hooks'
 
 export const useMerkleTreeWasmWorker = () => {
   const isWorkerReady = useWorker(MerkleTreeWorker)
